# internal/platform/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "小智服务端，包含OTA与Vision等接口",
        "title": "小智服务端 API 文档",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/admin/system": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取服务器的系统配置信息，包括选择的提供商和默认提示词",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "获取系统配置",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/webapi.SystemConfig"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "webapi.SystemConfig": {
            "type": "object",
            "properties": {
                "prompt": {
                    "type": "string"
                },
                "selectedASR": {
                    "type": "string"
                },
                "selectedLLM": {
                    "type": "string"
                },
                "selectedTTS": {
                    "type": "string"
                },
                "selectedVLLLM": {
                    "type": "string"
                }
            }
        }
    }
}